Preserve float temperature precision. Flat JSON for temperature and humidity with no sensor name prefix.

+++ sensorStation/esp32_sht30/main/SensorReading.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <variant>
 
-using SensorValue = std::variant<int, float>;
+using SensorValue = std::variant<float, int>;
 
 struct SensorReading
 {
